refine readme.md
