+++ ocr/armlinux/shell/cxx/ppocr_demo/build.sh
