+++ ocr/armlinux/shell/cxx/ppocr_demo/run.sh
