+++ ocr/armlinux/shell/cxx/ppocr_demo/src/CMakeLists.txt
