+++ ocr/armlinux/shell/cxx/ppocr_demo/src/db_post_process.h
@@ -14,9 +14,9 @@
 
 #pragma once
 
-#include <math.h>
 #include <iostream>
 #include <map>
+#include <math.h>
 #include <vector>
 
 #include "clipper.hpp"
