+++ ocr/armlinux/shell/cxx/ppocr_demo/src/pipeline.cc
@@ -140,20 +140,23 @@ cv::Mat Visualization(cv::Mat srcimg,
   }
 
   cv::imwrite(output_image_path, img_vis);
-  std::cout << "The detection visualized image saved in " << output_image_path.c_str() << std::endl;
+  std::cout << "The detection visualized image saved in "
+            << output_image_path.c_str() << std::endl;
   return img_vis;
 }
 
-Pipeline::Pipeline(const std::string &detModelDir, const std::string &clsModelDir,
-                    const std::string &recModelDir, const std::string &cPUPowerMode,
-                    const int cPUThreadNum,
-                    const std::string &config_path, const std::string &dict_path) {
-  clsPredictor_.reset(new ClsPredictor(
-      clsModelDir, cPUThreadNum, cPUPowerMode));
-  detPredictor_.reset(new DetPredictor(
-        detModelDir, cPUThreadNum, cPUPowerMode));
-  recPredictor_.reset(new RecPredictor(
-        recModelDir, cPUThreadNum, cPUPowerMode));
+Pipeline::Pipeline(const std::string &detModelDir,
+                   const std::string &clsModelDir,
+                   const std::string &recModelDir,
+                   const std::string &cPUPowerMode, const int cPUThreadNum,
+                   const std::string &config_path,
+                   const std::string &dict_path) {
+  clsPredictor_.reset(
+      new ClsPredictor(clsModelDir, cPUThreadNum, cPUPowerMode));
+  detPredictor_.reset(
+      new DetPredictor(detModelDir, cPUThreadNum, cPUPowerMode));
+  recPredictor_.reset(
+      new RecPredictor(recModelDir, cPUThreadNum, cPUPowerMode));
   Config_ = LoadConfigTxt(config_path);
   charactor_dict_ = ReadDict(dict_path);
   charactor_dict_.insert(charactor_dict_.begin(), "#"); // blank char for ctc
@@ -168,7 +171,8 @@ bool Pipeline::Process(std::string img_path, std::string output_img_path) {
   // Stage1: rec
   auto start = std::chrono::system_clock::now();
   // det predict
-  auto boxes = detPredictor_->Predict(srcimg, Config_, nullptr, nullptr, nullptr);
+  auto boxes =
+      detPredictor_->Predict(srcimg, Config_, nullptr, nullptr, nullptr);
 
   std::vector<float> mean = {0.5f, 0.5f, 0.5f};
   std::vector<float> scale = {1 / 0.5f, 1 / 0.5f, 1 / 0.5f};
@@ -182,15 +186,17 @@ bool Pipeline::Process(std::string img_path, std::string output_img_path) {
   for (int i = boxes.size() - 1; i >= 0; i--) {
     crop_img = GetRotateCropImage(img, boxes[i]);
     if (use_direction_classify >= 1) {
-      crop_img = clsPredictor_->Predict(crop_img, nullptr, nullptr, nullptr, 0.9);
+      crop_img =
+          clsPredictor_->Predict(crop_img, nullptr, nullptr, nullptr, 0.9);
     }
-    auto res = recPredictor_->Predict(crop_img, nullptr, nullptr, nullptr, charactor_dict_);
+    auto res = recPredictor_->Predict(crop_img, nullptr, nullptr, nullptr,
+                                      charactor_dict_);
     rec_text.push_back(res.first);
     rec_text_score.push_back(res.second);
   }
   auto end = std::chrono::system_clock::now();
   auto duration =
-  std::chrono::duration_cast<std::chrono::microseconds>(end - start);
+      std::chrono::duration_cast<std::chrono::microseconds>(end - start);
   //// visualization
   auto img_vis = Visualization(rgbaImage, boxes, output_img_path);
   // print recognized text
@@ -199,17 +205,18 @@ bool Pipeline::Process(std::string img_path, std::string output_img_path) {
               << std::endl;
   }
   std::cout << "花费了"
-          << double(duration.count()) *
+            << double(duration.count()) *
                    std::chrono::microseconds::period::num /
                    std::chrono::microseconds::period::den
-          << "秒" << std::endl;
+            << "秒" << std::endl;
   return true;
 }
 
 int main(int argc, char **argv) {
   if (argc < 7) {
     std::cerr << "[ERROR] usage: "
-              << " ./ocr_db_crnn_demo det_model_file cls_model_file rec_model_file image_path"
+              << " ./ocr_db_crnn_demo det_model_file cls_model_file "
+                 "rec_model_file image_path"
                  " output_image_path charactor_dict config\n";
     exit(1);
   }
@@ -222,8 +229,9 @@ int main(int argc, char **argv) {
   std::string config_path = argv[7];
   std::string cPUPowerMode = "";
   int cPUThreadNum = 1;
-  Pipeline* pipe = new Pipeline(det_model_file, cls_model_file, rec_model_file,
-                               cPUPowerMode, cPUThreadNum, config_path,  dict_path);
+  Pipeline *pipe =
+      new Pipeline(det_model_file, cls_model_file, rec_model_file, cPUPowerMode,
+                   cPUThreadNum, config_path, dict_path);
   pipe->Process(img_path, output_img_path);
   return 0;
 }
