+++ ocr/armlinux/shell/cxx/ppocr_demo/src/pipeline.h
@@ -14,9 +14,9 @@
 
 #pragma once
 #include "cls_process.h"
-#include "rec_process.h"
 #include "det_process.h"
 #include "paddle_api.h"
+#include "rec_process.h"
 #include <opencv2/core.hpp>
 #include <opencv2/highgui/highgui.hpp>
 #include <opencv2/imgcodecs.hpp>
@@ -29,11 +29,11 @@ class Pipeline {
 public:
   Pipeline(const std::string &detModelDir, const std::string &clsModelDir,
            const std::string &recModelDir, const std::string &cPUPowerMode,
-           const int cPUThreadNum,
-           const std::string &config_path, const std::string &dict_path);
+           const int cPUThreadNum, const std::string &config_path,
+           const std::string &dict_path);
 
   // bool Process(int inTextureId, int outTextureId, int textureWidth,
-              //  int textureHeight, std::string savedImagePath);
+  //  int textureHeight, std::string savedImagePath);
   bool Process(std::string img_path, std::string output_img_path);
 
 private:
