+++ ocr/armlinux/shell/cxx/ppocr_demo/src/rec_process.cc
@@ -56,7 +56,7 @@ RecPredictor::RecPredictor(const std::string &modelDir, const int cpuThreadNum,
 
 void RecPredictor::Preprocess(const cv::Mat &srcimg) {
   float wh_ratio =
-        static_cast<float>(srcimg.cols) / static_cast<float>(srcimg.rows);
+      static_cast<float>(srcimg.cols) / static_cast<float>(srcimg.rows);
   std::vector<float> mean = {0.5f, 0.5f, 0.5f};
   std::vector<float> scale = {1 / 0.5f, 1 / 0.5f, 1 / 0.5f};
   cv::Mat resize_img = CrnnResizeImg(srcimg, wh_ratio);
@@ -64,49 +64,50 @@ void RecPredictor::Preprocess(const cv::Mat &srcimg) {
 
   const float *dimg = reinterpret_cast<const float *>(resize_img.data);
 
-  std::unique_ptr<Tensor> input_tensor0(
-        std::move(predictor_->GetInput(0)));
+  std::unique_ptr<Tensor> input_tensor0(std::move(predictor_->GetInput(0)));
   input_tensor0->Resize({1, 3, resize_img.rows, resize_img.cols});
   auto *data0 = input_tensor0->mutable_data<float>();
   NHWC3ToNC3HW(dimg, data0, resize_img.rows * resize_img.cols, mean, scale);
 }
 
-std::pair<std::string, float> RecPredictor::Postprocess(const cv::Mat &rgbaImage,
-                    std::vector<std::string> charactor_dict) {
+std::pair<std::string, float>
+RecPredictor::Postprocess(const cv::Mat &rgbaImage,
+                          std::vector<std::string> charactor_dict) {
   // Get output and run postprocess
-    std::unique_ptr<const Tensor> output_tensor0(
-        std::move(predictor_->GetOutput(0)));
-    auto *predict_batch = output_tensor0->data<float>();
-    auto predict_shape = output_tensor0->shape();
+  std::unique_ptr<const Tensor> output_tensor0(
+      std::move(predictor_->GetOutput(0)));
+  auto *predict_batch = output_tensor0->data<float>();
+  auto predict_shape = output_tensor0->shape();
 
-    // ctc decode
-    std::string str_res;
-    int argmax_idx;
-    int last_index = 0;
-    float score = 0.f;
-    int count = 0;
-    float max_value = 0.0f;
+  // ctc decode
+  std::string str_res;
+  int argmax_idx;
+  int last_index = 0;
+  float score = 0.f;
+  int count = 0;
+  float max_value = 0.0f;
 
-    for (int n = 0; n < predict_shape[1]; n++) {
-      argmax_idx = int(Argmax(&predict_batch[n * predict_shape[2]],
-                              &predict_batch[(n + 1) * predict_shape[2]]));
-      max_value =
-          float(*std::max_element(&predict_batch[n * predict_shape[2]],
-                                  &predict_batch[(n + 1) * predict_shape[2]]));
-      if (argmax_idx > 0 && (!(n > 0 && argmax_idx == last_index))) {
-        score += max_value;
-        count += 1;
-        str_res += charactor_dict[argmax_idx];
-      }
-      last_index = argmax_idx;
+  for (int n = 0; n < predict_shape[1]; n++) {
+    argmax_idx = int(Argmax(&predict_batch[n * predict_shape[2]],
+                            &predict_batch[(n + 1) * predict_shape[2]]));
+    max_value =
+        float(*std::max_element(&predict_batch[n * predict_shape[2]],
+                                &predict_batch[(n + 1) * predict_shape[2]]));
+    if (argmax_idx > 0 && (!(n > 0 && argmax_idx == last_index))) {
+      score += max_value;
+      count += 1;
+      str_res += charactor_dict[argmax_idx];
     }
-    score /= count;
-    return std::make_pair(str_res, score);
-}  
+    last_index = argmax_idx;
+  }
+  score /= count;
+  return std::make_pair(str_res, score);
+}
 
-std::pair<std::string, float> RecPredictor::Predict(const cv::Mat &rgbaImage,
-                           double *preprocessTime, double *predictTime,
-                           double *postprocessTime, std::vector<std::string> charactor_dict) {
+std::pair<std::string, float>
+RecPredictor::Predict(const cv::Mat &rgbaImage, double *preprocessTime,
+                      double *predictTime, double *postprocessTime,
+                      std::vector<std::string> charactor_dict) {
   // auto t = GetCurrentTime();
   Preprocess(rgbaImage);
   // *preprocessTime = GetElapsedTime(t);
