+++ ocr/armlinux/shell/cxx/ppocr_demo/src/rec_process.h
@@ -14,11 +14,11 @@
 
 #pragma once
 
-#include "utils.h"
 #include "opencv2/core.hpp"
 #include "opencv2/imgcodecs.hpp"
 #include "opencv2/imgproc.hpp"
 #include "paddle_api.h"
+#include "utils.h"
 using namespace paddle::lite_api; // NOLINT
 
 class RecPredictor {
@@ -26,15 +26,16 @@ public:
   explicit RecPredictor(const std::string &modelDir, const int cpuThreadNum,
                         const std::string &cpuPowerMode);
 
-  std::pair<std::string, float> Predict(const cv::Mat &rgbaImage, double *preprocessTime,
-                           double *predictTime, double *postprocessTime, std::vector<std::string> charactor_dict);
+  std::pair<std::string, float>
+  Predict(const cv::Mat &rgbaImage, double *preprocessTime, double *predictTime,
+          double *postprocessTime, std::vector<std::string> charactor_dict);
 
 private:
   void Preprocess(const cv::Mat &rgbaImage);
-  std::pair<std::string, float> Postprocess(const cv::Mat &rgbaImage, std::vector<std::string> charactor_dict);
+  std::pair<std::string, float>
+  Postprocess(const cv::Mat &rgbaImage,
+              std::vector<std::string> charactor_dict);
 
 private:
   std::shared_ptr<paddle::lite_api::PaddlePredictor> predictor_;
 };
-
-
