+++ ocr/armlinux/shell/cxx/ppocr_demo/src/utils.cc
